excluir categoria

+++ codigo/deletarcategoria.php
@@ -1,6 +1,7 @@
 <?php
-    require_once "conexao.php";
+require_once "conexao.php";
 
+if (isset($_GET['idcategoria'])) {
     $id = $_GET['idcategoria'];
 
     $sql = "DELETE FROM tb_categorias WHERE idcategoria = $id";
@@ -8,4 +9,9 @@
     mysqli_query($conexao, $sql);
 
     header("Location: listarcategoria.php");
+    exit;
+} else {
+    echo "idcategoria não fornecido.";
+    exit;
+}
 ?>
